// src/common/interfaces/users/UserConsts.tsx
/* eslint-disable no-unused-vars */
export enum SettingKey {
    USER_UID = 'useruid',
    CREATED = 'created',
    UPDATED = 'updated',
    STOCKNUM_PREFIX = 'stocknumPrefix',
    STOCKNUM_SUFFIX = 'stocknumSuffix',
    STOCKNUM_FIXED_DIGITS = 'stocknumFixedDigits',
    STOCKNUM_SEQUENTIAL = 'stocknumSequental',
    STOCKNUMTI_SEQUENTIAL = 'stocknumtiSequental',
    STOCKNUMTI_FROM_SOLD_VEHICLE = 'stocknumtiFromSoldVehicle',
    STOCKNUM_LAST_6_OF_VIN = 'stocknumLast6ofVIN',
    STOCKNUM_LAST_8_OF_VIN = 'stocknumLast8ofVIN',
    DEAL_TYPE = 'dealType',
    DEAL_STATUS = 'dealStatus',
    LEASE_TERM = 'leaseTerm',
    LEASE_PAYMENT_FREQUENCY = 'leasePaymentFrequency',
    INVENTORY_STATUS = 'inventoryStatus',
    SALE_TYPE = 'saleType',
    ACCOUNT_FIXED_DIGITS = 'accountFixedDigits',
    ACCOUNT_LATE_FEE_GRACE_PERIOD = 'accountLateFeeGracePeriod',
    ACCOUNT_LATE_FEE_MAX = 'accountLateFeeMax',
    ACCOUNT_LATE_FEE_MIN = 'accountLateFeeMin',
    ACCOUNT_LATE_FEE_PERCENTAGE = 'accountLateFeePercentage',
    ACCOUNT_PREFIX = 'accountPrefix',
    ACCOUNT_START_NUMBER = 'accountStartNumber',
    ACCOUNT_SUFFIX = 'accountSuffix',
    CONTRACT_DEF_INTEREST_RATE = 'contractDefInterestRate',
    CONTRACT_PAYMENT_FREQUENCY = 'contractPaymentFrequency',
    FEE_DEF_DOCUMENTATION = 'feeDefDocumentation',
    FEE_DEF_SPARE_TAG = 'feeDefSpareTag',
    FEE_DEF_SPARE_TRANSFER_TAG = 'feeDefSpareTransferTag',
    FEE_DEF_TAG = 'feeDefTag',
    FEE_DEF_TITLE = 'feeDefTitle',
    FEE_DEF_TRANSFER = 'feeDefTransfer',
    FEE_DEF_VEHICLE_PACK = 'feeDefvehiclePack',
    INDEX = 'index',
    ITEM_UID = 'itemuid',
    LEASE_DEFAULT_MILEAGE = 'leaseDefaultMileage',
    LEASE_MONEY_FACTOR = 'leaseMoneyFactor',
    LEASE_OVERAGE_AMOUNT = 'leaseOverageAmount',
    TAX_DEF_STATE_VEHICLE_TAX_RATE = 'taxDefStateVehicleTaxRate',
}